fix(ui): fix user component
remove junk, revise component structure
