+++ src/app/app.module.ts
@@ -36,6 +36,7 @@ import { DnbhubPlaylistsComponent } from './components/playlists/playlists.compo
 import { DnbhubProgressBarComponent } from './components/progress-bar/progress-bar.component';
 import { DnbhubToolbarComponent } from './components/toolbar/toolbar.component';
 import { DnbhubUserMeComponent } from './components/user-me/user-me.component';
+import { DnbhubUserPlaylistsComponent } from './components/user-playlists/user-playlists.component';
 import { DnbhubAboutStoreModule } from './state/about/about.module';
 import { DnbhubAdminStoreModule } from './state/admin/admin.module';
 import { DnbhubBlogStoreModule } from './state/blog/blog.module';
@@ -74,6 +75,7 @@ const ngxsLoggerPluginOptions: NgxsLoggerPluginOptions = {
     DnbhubAboutComponent,
     DnbhubUserComponent,
     DnbhubUserMeComponent,
+    DnbhubUserPlaylistsComponent,
     DnbhubAdminComponent,
     DnbhubSoundcloudPlayerComponent,
     DnbhubContactDialogComponent,
