+++ src/app/components/user-playlists/user-playlists.component.scss
@@ -1,0 +1,6 @@
+:host {
+  display: block;
+  width: 100%;
+  height: auto !important;
+  overflow-y: auto;
+}
